+++ src/treetop/main.ts
@@ -5,13 +5,9 @@ import '@fontsource/inter/500.css';
 import '@fontsource/material-icons';
 import '@fontsource/news-cycle/latin-400.css';
 import 'normalize.css/normalize.css';
+import './main.css';
 
 import Treetop from './Treetop.svelte';
-
-// Import after Treetop component to ensure that MDC style overrides follow
-// MDC/SMUI rules in the bundle.
-// See https://github.com/hperrin/svelte-material-ui/issues/129
-import './main.css';
 
 // Get root bookmark ID from hash
 const rootBookmarkId = decodeURIComponent(location.hash.substr(1));